Fixing Power LED
Right now the 'Power LED' is run from Ground to... Ground, so it's
not exactly going to work 'right'.  When moving things around in
Eagle it didn't attach properly to the 3.3V line on the schematic
and so was left dangling, despite it being connected.

This, thankfully, doesn't affect the boards except cause a little
bit of build waste.  This is fixed now, though it'll likely get
set to DNI, as it's only really useful on dev boards.